adding generate_labelled_commits.sh
According to linux's devel git repository and stable repository, the commit label is automatically generated.

For example, a commit in the Linux devel repository:
Commit b248371628aad599a48540962f6b85a21a8a0c3f
Author: Takashi Iwai <tiwai@suse.de>
Date: Sun Jan 31 10:32:37 2016 +0100

	ALSA: pcm: Fix potential deadlock in OSS emulation

It corresponds to the commit in the origin/linux-2.6.33.y branch of the stable repository:
Commit 937524d03c6eaf3a5d00b714eaaf15a1a9c8478a
Author: Takashi Iwai <tiwai@suse.de>
Date: Sun Jan 31 10:32:37 2016 +0100

	ALSA: pcm: Fix potential deadlock in OSS emulation

	Commit b248371628aad599a48540962f6b85a21a8a0c3f upstream.

We based on the keywords in the commit log of a branch in the stable repository:
Commit b248371628aad599a48540962f6b85a21a8a0c3f upstream,
to determine whether it should be marked with a stable label.